Modifications de Zombies

+++ src/main/java/com/epf/API/dto/ZombiesDto.java
@@ -3,18 +3,19 @@ package com.epf.API.dto;
 public class ZombiesDto {
     private int id_zombie;
     private String nom;
-    private Integer point_de_vie;
-    private Integer degat_attaque;
-    private int id_map;
-    private Integer attaque_par_seconde;
-    private Integer vitesse_de_deplacement;
+    private int point_de_vie;        
+    private int degat_attaque;         
+    private Integer id_map;           
+    private Double attaque_par_seconde;    
+    private Double vitesse_de_deplacement; 
     private String chemin_image;
 
     public ZombiesDto() {
     }
 
-    public ZombiesDto(int id_zombie, String nom, Integer point_de_vie, Integer degat_attaque,
-                      int id_map, Integer attaque_par_seconde, Integer vitesse_de_deplacement, String chemin_image) {
+    public ZombiesDto(int id_zombie, String nom, int point_de_vie, int degat_attaque,
+                      Integer id_map, Double attaque_par_seconde, 
+                      Double vitesse_de_deplacement, String chemin_image) {
         this.id_zombie = id_zombie;
         this.nom = nom;
         this.point_de_vie = point_de_vie;
@@ -41,43 +42,43 @@ public class ZombiesDto {
         this.nom = nom;
     }
 
-    public Integer getPoint_de_vie() {
+    public int getPoint_de_vie() {
         return point_de_vie;
     }
 
-    public void setPoint_de_vie(Integer point_de_vie) {
+    public void setPoint_de_vie(int point_de_vie) {
         this.point_de_vie = point_de_vie;
     }
 
-    public Integer getDegat_attaque() {
+    public int getDegat_attaque() {
         return degat_attaque;
     }
 
-    public void setDegat_attaque(Integer degat_attaque) {
+    public void setDegat_attaque(int degat_attaque) {
         this.degat_attaque = degat_attaque;
     }
 
-    public int getId_map() {
+    public Integer getId_map() {
         return id_map;
     }
 
-    public void setId_map(int id_map) {
+    public void setId_map(Integer id_map) {
         this.id_map = id_map;
     }
 
-    public Integer getAttaque_par_seconde() {
+    public Double getAttaque_par_seconde() {
         return attaque_par_seconde;
     }
 
-    public void setAttaque_par_seconde(Integer attaque_par_seconde) {
+    public void setAttaque_par_seconde(Double attaque_par_seconde) {
         this.attaque_par_seconde = attaque_par_seconde;
     }
 
-    public Integer getVitesse_de_deplacement() {
+    public Double getVitesse_de_deplacement() {
         return vitesse_de_deplacement;
     }
 
-    public void setVitesse_de_deplacement(Integer vitesse_de_deplacement) {
+    public void setVitesse_de_deplacement(Double vitesse_de_deplacement) {
         this.vitesse_de_deplacement = vitesse_de_deplacement;
     }
